COM-3946 - can't mutualize in composable because of refreshing datas
